Timeout can be now set in the Resolver.__init__().
User module can now set the required timeout. During testing this
feature I found out, that for very short times (0.001 sec), a "Network
is unreachable" socket error (errno 101) appears instead of
socket.timeout. But logically, this is a timeout as well.

Therefore I decided to catch this exception and treat it the very same
way I treat the socket.timeout one. This is handled in the _http module.