Method Resolver._store_into_db() now calls save_header() from HttpHeaderMeta.
Yet misses saving the content and proper connection via the content_id
field. The change was not properly tested yet since I have neither
implementation of the database model nor a running MongoDB.

Module model.py was changed only in the HttpHeaderMeta
.save_header() in order to reflect that headers are passed as dictionary
from the _http module which is it puts [] in proper places.